Batched UI updates and guard against dispatch in reducers
Yet another React feature that's actually just a necessity for working with a virtual
DOM, not a feature enabled by it.

The update step really does have to be asynchronous and entirely distinct from state
updates, because mounting and unmouting elements can trigger state changes, even with
stateless view functions: a focused `input` being removed form the DOM triggers a `blur`
event.

Unfortunately, brute-forcing a full DOM update in a tight requestAnimationFrame loop
pegs my CPU at 20%. And it's the loop itself, not the DOM updates that is expensive --
adding logic to see if state actually changed or not doesn't bring the CPU down.

The logic to only request an animation frame after state has changed, and to avoid
requesting more frames until the last request completed, is sucky. Fine and doable, but
sucky.

Recoil tries to be as simple and transparent as possible and as correct as possible,
with small regard for efficiency, so I'm tempted to leave it as-is, but pegging CPU when
nothing is happening seems almost too egregious even for recoil. So I'll probably eat
the 10 lines and extra complexity and implement it.